# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "money",
    "description": "Budgeting and shared-expense tracking, API first, git-backed.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/budgets/{budget}/entries/{entry_id}/history": {
      "get": {
        "tags": [
          "entries"
        ],
        "summary": "Commits that touched this entry",
        "operationId": "getEntryHistory",
        "parameters": [
          {
            "name": "entry_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Entry Id"
            }
          },
          {
            "name": "budget",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Budget"
            }
          },
          {
            "name": "X-Act-As",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Act-As"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "money_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Money Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-cli": {
          "command": "entry history",
          "args": [
            "entry_id"
          ]
        }
      }
    },
...
    "/api/v1/budgets/{budget}/history": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "Everything that has changed, newest first",
        "description": "One feed for every change, whichever file it touched.\n\nAssignments show up next to entries because they are the same kind of event to the person\nreading \u2014 money moved \u2014 even though one lives in a ledger file and the other in a map.",
        "operationId": "getHistory",
        "parameters": [
          {
            "name": "budget",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Budget"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "X-Act-As",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Act-As"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "money_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Money Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-cli": {
          "command": "history"
        }
      }
    },
    "/api/v1/budgets/{budget}/history/{sha}": {
      "get": {
        "tags": [
          "history"
        ],
        "summary": "What one change actually did",
        "description": "The diff, so \"what did this change\" has an answer that is not a guess.\n\nA subject line names what someone meant to do. The patch is what actually happened, and\nfor a data model stored as YAML it reads well enough to be the answer rather than a\ndebugging aid \u2014 an assignment is one number changing on one line.\n\nWhether it can still be undone is answered here rather than left to the client, because\nit needs the whole log: a change that something later reverted is no longer applied, and\noffering to undo it again would revert the revert.",
        "operationId": "getHistoryDetail",
        "parameters": [
          {
            "name": "sha",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sha"
            }
          },
          {
            "name": "budget",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Budget"
            }
          },
          {
            "name": "X-Act-As",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Act-As"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "money_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Money Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoryDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-cli": {
          "command": "history show",
          "args": [
            "sha"
          ]
        }
      }
    },
    "/api/v1/budgets/{budget}/history/undo": {
      "post": {
        "tags": [
          "history"
        ],
        "summary": "Undo a change you made",
        "description": "Revert a change, defaulting to your most recent un-reverted one.\n\nScoped to your own changes: in a shared budget, silently reversing your partner's work\nfrom your phone is the failure worth designing against. Naming a sha reverts anyone's,\nwhich makes doing so deliberate.",
        "operationId": "undoChange",
        "parameters": [
          {
            "name": "budget",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Budget"
            }
          },
          {
            "name": "sha",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sha"
            }
          },
          {
            "name": "X-Act-As",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Act-As"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "money_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Money Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UndoResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-cli": {
          "command": "undo"
        }
      }
    },
    "/api/v1/budgets/{budget}/history/redo": {
      "post": {
        "tags": [
          "history"
        ],
        "summary": "Re-apply something you undid",
        "description": "Re-apply the change your last undo reversed, by reverting the revert.\n\nOnly available while the undo is still your most recent change. Anything done afterwards\nclears it, the way a new edit clears an editor's redo stack \u2014 re-applying a change on top\nof later work would produce a state nobody asked for.",
        "operationId": "redoChange",
        "parameters": [
          {
            "name": "budget",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Budget"
            }
          },
          {
            "name": "X-Act-As",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Act-As"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          },
          {
            "name": "money_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Money Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UndoResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-cli": {
          "command": "redo"
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommitRef": {
        "properties": {
          "sha": {
            "type": "string",
            "title": "Sha"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "author_name": {
            "type": "string",
            "title": "Author Name"
          },
          "date": {
            "type": "string",
            "title": "Date"
          }
        },
        "type": "object",
        "required": [
          "sha",
          "subject",
          "author_name",
          "date"
        ],
        "title": "CommitRef"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HistoryDetail": {
        "properties": {
          "sha": {
            "type": "string",
            "title": "Sha"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "entry_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Id"
          },
          "person": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Person",
            "description": "Whose buckets or assignments moved"
          },
          "month": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Month"
          },
          "mine": {
            "type": "boolean",
            "title": "Mine"
          },
          "reverts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reverts",
            "description": "Sha this change undid, if any"
          },
          "reverted_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reverted By",
            "description": "Sha that undid this change; set means it is no longer applied"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/HistoryFileChange"
            },
            "type": "array",
            "title": "Files"
          },
          "diff": {
            "type": "string",
            "title": "Diff",
            "description": "Unified patch, possibly truncated"
          },
          "diff_truncated": {
            "type": "boolean",
            "title": "Diff Truncated"
          },
          "can_undo": {
            "type": "boolean",
            "title": "Can Undo",
            "description": "False when it is already undone, or you have read-only access"
          }
        },
        "type": "object",
        "required": [
          "sha",
          "subject",
          "kind",
          "author",
          "actor",
          "date",
          "entry_id",
          "mine",
          "files",
          "diff",
          "diff_truncated",
          "can_undo"
        ],
        "title": "HistoryDetail",
        "description": "One change, in enough detail to decide whether to undo it."
      },
      "HistoryEvent": {
        "properties": {
          "sha": {
            "type": "string",
            "title": "Sha"
          },
          "subject": {
            "type": "string",
            "title": "Subject"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "entry, assignment, bucket, rules, members, note, undo"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "actor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Actor"
          },
          "date": {
            "type": "string",
            "title": "Date"
          },
          "entry_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entry Id"
          },
          "mine": {
            "type": "boolean",
            "title": "Mine",
            "description": "You made this change"
          },
          "reverts": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reverts",
            "description": "Sha this change undid, if any"
          }
        },
        "type": "object",
        "required": [
          "sha",
          "subject",
          "kind",
          "author",
          "actor",
          "date",
          "entry_id",
          "mine"
        ],
        "title": "HistoryEvent"
      },
      "HistoryFileChange": {
        "properties": {
          "path": {
            "type": "string",
            "title": "Path"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "added, modified or removed"
          },
          "added": {
            "type": "integer",
            "title": "Added",
            "description": "Lines added"
          },
          "removed": {
            "type": "integer",
            "title": "Removed",
            "description": "Lines removed"
          }
        },
        "type": "object",
        "required": [
          "path",
          "status",
          "added",
          "removed"
        ],
        "title": "HistoryFileChange"
      },
      "HistoryList": {
        "properties": {
          "commits": {
            "items": {
              "$ref": "#/components/schemas/CommitRef"
            },
            "type": "array",
            "title": "Commits"
          }
        },
        "type": "object",
        "required": [
          "commits"
        ],
        "title": "HistoryList"
      },
      "HistoryPage": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/HistoryEvent"
            },
            "type": "array",
            "title": "Events"
          },
          "undoable": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Undoable",
            "description": "What undo would revert, if anything"
          },
          "redoable": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Redoable",
            "description": "Set only while your last change was an undo"
          }
        },
        "type": "object",
        "required": [
          "events",
          "undoable"
        ],
        "title": "HistoryPage"
      },
...
      "UndoResult": {
        "properties": {
          "undone_sha": {
            "type": "string",
            "title": "Undone Sha"
          },
          "undone_subject": {
            "type": "string",
            "title": "Undone Subject",
            "description": "What was reverted, for the confirmation"
          },
          "commit": {
            "type": "string",
            "title": "Commit"
          },
          "can_redo": {
            "type": "boolean",
            "title": "Can Redo"
          }
        },
        "type": "object",
        "required": [
          "undone_sha",
          "undone_subject",
          "commit",
          "can_redo"
        ],
        "title": "UndoResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}